refactor(App.tsx): move AddUser component to a separate folder and update import statement
refactor(Navigation.tsx): add conditional rendering to the footer to hide it on the AddUser page

feat(AddUserNAV.tsx): create AddUserButtonNavigation component with buttons to submit form and clear fields, hides the "main" navigation footer

feat(apiContextProvider.tsx): update getUsers function to fetch all users from the API, Api link updated
