+++ src/App.tsx
@@ -7,7 +7,7 @@ import { useIsAuthenticated } from '@azure/msal-react'
 import { LandingPage } from './pages/landingPage/LandingPage'
 import { Login } from './pages/login'
 import { Profile } from './pages/profile'
-import { AddUser } from './pages/users/AddUser'
+import { AddUser } from './pages/users/addUser/AddUser'
 import { ApiContextProvider } from './pages/users/context/apiContextProvider'
 import { AuthProvider } from './pages/landingPage/context/LandingPageContextProvider'
 import { ListUsers } from './pages/users/listUsers/ListUsers'
